RAG/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.41.0
sqlalchemy>=2.0.0
psycopg2-binary
python-dotenv
tqdm
fastapi
uvicorn[standard]
